+++ R/zzz.R
@@ -1,11 +1,7 @@
 .onAttach <- function(libname, pkgname) {
   RFpredInterval.version <- read.dcf(file=system.file("DESCRIPTION", package=pkgname),
                             fields="Version")
-  packageStartupMessage(paste("\n",
-                              pkgname,
+  packageStartupMessage(paste(pkgname,
                               RFpredInterval.version,
-                              "\n",
-                              "\n",
-                              "\n",
                               "\n"))
 }