+++ NEWS.md
@@ -1,3 +1,9 @@
+## RFpredInterval 1.0.5
+* Added a new argument to `pibf()` and `rfpi()` functions for returning out-of-bag (OOB) predictions and prediction intervals.
+* Updated `pibf()`, `rfpi()` and `piall()` output classes to enable a package print method.
+* Added a package print method.
+* Changed plot method to a S3 method.
+
 ## RFpredInterval 1.0.4
 Fixed the omp.h declaration before R headers per Prof. Ripley, in anticipation of clang 13.0.0.